refactor: input -> numUtil 배열 변환 책임 전가
