+++ src/main/java/calculator/input/Input.java
@@ -1,7 +1,5 @@
 package calculator.input;
 
-import calculator.Utils.NumUtil;
-
 public class Input {
 
     private static final char SEPARATOR_COMMA = ',';
@@ -17,27 +15,11 @@ public class Input {
         return InputFilter.from(input);
     }
 
-    public int[] getNums() {
-        int[] numArr = new int[nums.length() + 1];
-        int idx = 0, cur = 0;
-        boolean isPrevSeparator = false;
-        for (int i = 0; i < nums.length(); i++) {
-            char ch = nums.charAt(i);
-            if (isSeparator(ch) && !isPrevSeparator) {
-                numArr[idx++] = cur;
-                cur = 0;
-                isPrevSeparator = true;
-                continue;
-            }
-            cur *= 10;
-            cur += NumUtil.toInt(ch);
-            isPrevSeparator = false;
-        }
-        numArr[idx] = cur;
-        return numArr;
+    public boolean isSeparatorEqual(char ch) {
+        return ch == SEPARATOR_COLON || ch == SEPARATOR_COMMA;
     }
 
-    protected boolean isSeparator(char ch) {
-        return ch == SEPARATOR_COLON || ch == SEPARATOR_COMMA;
+    public String getNums() {
+        return nums;
     }
 }
